(Mac) Makefile changes for supporting Metal build (#11929)
* (QB) Add rule for Metal - define HAVE_COCOA_METAL for now
when building for Metal, and HAVE_COCOA for non-Metal build

* Make necessary changes to Makefile.common and qb/config.libs.sh:
- Disable OpenGL for now for Metal build
- Take Metal conditional out of OpenGL block in Makefile.common

* (Metal) Header fixes for compiling without Xcode/Griffin

* (Mac) Some buildfixes for non-Xcode building

* (cocoa_common.h) Restructure to no longer use -DOSX

* (Apple) More buildfixes

* (Apple) Cleanups

+++ ui/drivers/ui_cocoa.m
@@ -121,7 +121,7 @@ static char **waiting_argv;
             old_flags                             = new_flags;
 
             apple_input_keyboard_event(down, keycode,
-                  0, new_flags, RETRO_DEVICE_KEYBOARD);
+                  0, (uint32_t)new_flags, RETRO_DEVICE_KEYBOARD);
          }
          break;
         case NSEventTypeMouseMoved:
